requirements.txt
b2sdk
gitignore_parser
typer
pyyaml
